requirements.txt
requests
python-dotenv
pymongo
notebook
pandas
azure-core
azure-identity
azure-search-documents
pypdf
docx2txt
faiss-cpu
streamlit
streamlit-chat
langchain
langchain-community
langchain-experimental
langchain-core
huggingface-hub
nltk
openai
langchain-openai
streamlit-navigation-bar
pymupdf
bcrypt 
plotly
msal
google-cloud-storage
PyPDF2
streamlit_elements





